requirements.txt
# HealthGuard — Dependencies
openai>=1.12.0
httpx>=0.27.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
structlog>=24.1.0
python-dotenv>=1.0.0
cryptography>=42.0.0
Pillow>=10.2.0
python-multipart>=0.0.6
